+++ src/com/intellij/idea/plugin/hybris/acl/formatting/AclFormattingModelBuilder.kt
@@ -1,0 +1,39 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.intellij.idea.plugin.hybris.acl.formatting
+
+import com.intellij.formatting.FormattingContext
+import com.intellij.formatting.FormattingModel
+import com.intellij.formatting.FormattingModelBuilder
+import com.intellij.formatting.FormattingModelProvider
+
+class AclFormattingModelBuilder : FormattingModelBuilder {
+
+    override fun createModel(formattingContext: FormattingContext): FormattingModel {
+        val element = formattingContext.psiElement
+        val settings = formattingContext.codeStyleSettings
+        val rootBlock = AclBlock(
+            node = element.node,
+            spacingBuilder = AclSpacingBuilder(settings, settings.getCustomSettings(AclCodeStyleSettings::class.java)),
+            codeStyleSettings = settings,
+            alignmentStrategy = AclAlignmentStrategy()
+        )
+
+        return FormattingModelProvider.createFormattingModelForPsiFile(element.containingFile, rootBlock, settings)
+    }
+}